Cntd. fault tolerant N-Triples/N-Quads reader
